# openapi.yaml
openapi: 3.0.3
info:
  title: Booking Service API
  version: "1.0"
  description: >
    A minimal booking-domain API, versioned under /v1. Designed to be
    consumable by both conventional clients and AI-agent/LLM-driven clients:
    every error response carries a stable machine-readable `code`, and a
    dedicated /summary endpoint returns pre-computed, natural-language-ready
    fields so an agent doesn't need to do date arithmetic itself.
servers:
  - url: http://localhost:8080
paths:
  /v1/bookings:
    post:
      summary: Create a booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
    get:
      summary: List all bookings
      operationId: listBookings
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /v1/bookings/{id}:
    get:
      summary: Get a booking by ID
      operationId: getBooking
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: The booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/Error'
  /v1/bookings/{id}/summary:
    get:
      summary: Get an agent-friendly summary of a booking
      description: >
        Returns pre-computed derived fields (night count, human-readable
        summary) so LLM-driven clients don't need to compute date arithmetic
        themselves to reason about a booking.
      operationId: getBookingSummary
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummary'
        '404':
          $ref: '#/components/responses/Error'
  /v1/bookings/{id}/confirm:
    post:
      summary: Confirm a pending booking
      operationId: confirmBooking
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/bookings/{id}/cancel:
    post:
      summary: Cancel a booking
      operationId: cancelBooking
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
components:
  parameters:
    BookingId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Structured error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateBookingRequest:
      type: object
      required: [guest_name, property_id, check_in, check_out]
      properties:
        guest_name:
          type: string
        property_id:
          type: string
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
    Booking:
      type: object
      properties:
        id:
          type: string
        guest_name:
          type: string
        property_id:
          type: string
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingSummary:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        nights:
          type: integer
        human_summary:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable, machine-readable error identifier
            message:
              type: string
              description: Human-readable explanation